peptide transposition v 1.1.0 - can now only consider slices of the target protein, important(?) for polyprotein analysis

+++ peptide_mapping.py
@@ -299,8 +299,9 @@ class PeptideMapping(object):
 
     @staticmethod
     def calc_coord_diff(trgt_match, query_from, query_to):
-        return (trgt_match.get_target_start() + (trgt_match.get_target_end() - trgt_match.get_target_start()) / 2) - (
-                query_from + (query_to - query_from) / 2)
+        return int(
+            (trgt_match.get_target_start() + (trgt_match.get_target_end() - trgt_match.get_target_start()) / 2) - (
+                    query_from + (query_to - query_from) / 2))
 
     @staticmethod
     def write_header(of):
